move OpenChargeMap API's network layer to a separate package
